Authentication lockdown.

+++ core/views.py
@@ -1,4 +1,9 @@
 from django.http.response import HttpResponseBadRequest, HttpResponseForbidden
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.mixins import LoginRequiredMixin
+from django.http import JsonResponse
+from django.views.decorators.csrf import ensure_csrf_cookie
+from django.views.decorators.http import require_POST
 from django.shortcuts import render, get_object_or_404
 from django.core.paginator import Paginator
 from django.views import View
@@ -12,11 +17,14 @@ from rest_framework import viewsets, status, mixins
 from rest_framework.exceptions import MethodNotAllowed
 from rest_framework.decorators import action
 from rest_framework.response import Response
+
 from core.models import Affiliations, Contact, Experience, FooterStat, Project, Skills
 from core.serializers import (
     AffiliationSerializer, ContactSerializer, ExperienceSerializer, ExperienceCreationSerializer,
     FooterStatSerializer, ProjectSerializer, SkillsSerializer
 )
+
+import json
 import re
 
 
@@ -100,9 +108,8 @@ class SkillViewSet(
         })
 
     def patch(self, request, *args, **kwargs):
-        # TODO protect behind authentication
-        # if not request.user.is_authenticated:
-        #     return Response({'message': 'Forbidden'}, status=status.HTTP_403_FORBIDDEN)
+        if not request.user.is_authenticated:
+            return Response({'message': 'Forbidden'}, status=status.HTTP_403_FORBIDDEN)
         item = self.get_queryset()
         serializer = self.get_serializer(data=request.data, instance=item)
         if not serializer.is_valid():
@@ -125,7 +132,8 @@ class ProjectViewSet(
 
     @action(detail=False, methods=['PATCH'], name='Update a group of projects at once')
     def batch_update(self, request):
-        # TODO: protect behind authentication
+        if not request.user.is_authenticated:
+            return Response({'message': 'Forbidden'}, status=status.HTTP_403_FORBIDDEN)
         validated_serializers = []
         for item in request.data:
             try:
@@ -150,13 +158,15 @@ class ProjectViewSet(
         raise MethodNotAllowed('PUT', detail='Use Patch')
 
     def destroy(self, request, pk):
-        # TODO: protect behind authentication
+        if not request.user.is_authenticated:
+            return Response({'message': 'Forbidden'}, status=status.HTTP_403_FORBIDDEN)
         project = get_object_or_404(Project, pk=pk)
         project.delete()
         return Response({'message': 'Success'})
 
     def create(self, request, *args, **kwargs):
-        # TODO: protect behind authentication
+        if not request.user.is_authenticated:
+            return Response({'message': 'Forbidden'}, status=status.HTTP_403_FORBIDDEN)
         serializer = self.get_serializer(data=request.data)
         if not serializer.is_valid():
             return Response({
@@ -167,7 +177,8 @@ class ProjectViewSet(
         return Response({'message': "Success", 'project': self.get_serializer(project).data})
 
     def patch(self, request, pk,):
-        # TODO: protect behind authentication
+        if not request.user.is_authenticated:
+            return Response({'message': 'Forbidden'}, status=status.HTTP_403_FORBIDDEN)
         project = self.get_object(pk)
         serializer = self.get_serializer(project, data=request.data)
         if not serializer.is_valid():
@@ -247,13 +258,15 @@ class AffiliationViewSet(
         raise MethodNotAllowed('PUT', detail='Use Patch')
 
     def destroy(self, request, pk):
-        # TODO: protect behind authentication
+        if not request.user.is_authenticated:
+            return Response({'message': 'Forbidden'}, status=status.HTTP_403_FORBIDDEN)
         project = get_object_or_404(Affiliations, pk=pk)
         project.delete()
         return Response({'message': 'Success'})
 
     def create(self, request, *args, **kwargs):
-        # TODO: protect behind authentication
+        if not request.user.is_authenticated:
+            return Response({'message': 'Forbidden'}, status=status.HTTP_403_FORBIDDEN)
         serializer = self.get_serializer(data=request.data)
         if not serializer.is_valid():
             return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
@@ -261,7 +274,8 @@ class AffiliationViewSet(
         return Response(self.get_serializer(project).data)
 
     def patch(self, request, pk,):
-        # TODO: protect behind authentication
+        if not request.user.is_authenticated:
+            return Response({'message': 'Forbidden'}, status=status.HTTP_403_FORBIDDEN)
         affiliation = get_object_or_404(Affiliations, pk=pk)
         serializer = self.get_serializer(affiliation, data=request.data)
         if not serializer.is_valid():
@@ -305,9 +319,9 @@ class ExperienceViewSet(
         return Response(serializer.data)
 
     def patch(self, request, pk,):
-        # TODO: protect behind authentication
+        if not request.user.is_authenticated:
+            return Response({'message': 'Forbidden'}, status=status.HTTP_403_FORBIDDEN)
         project = get_object_or_404(self.model, pk=pk)
-        print(request.data)
         serializer = self.creation_serializer(project, data=request.data)
         if not serializer.is_valid():
             return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
@@ -318,15 +332,16 @@ class ExperienceViewSet(
         raise MethodNotAllowed('PUT', detail='Use Patch')
 
     def destroy(self, request, pk):
-        # TODO: protect behind authentication
+        if not request.user.is_authenticated:
+            return Response({'message': 'Forbidden'}, status=status.HTTP_403_FORBIDDEN)
         experience = get_object_or_404(self.model, pk=pk)
         experience.delete()
         return Response({'message': 'Success'})
 
     def create(self, request):
-        # TODO: Protect behind authentication and authorization
+        if not request.user.is_authenticated:
+            return Response({'message': 'Forbidden'}, status=status.HTTP_403_FORBIDDEN)
         data = request.data
-        print(data)
         serializer = self.creation_serializer(data=data)
         if serializer.is_valid():
             exp = serializer.save()
@@ -338,3 +353,45 @@ class Client(View):
 
     def get(self, request):
         return render(request, 'build/index.html', {})
+
+
+class AuthenticatedClient(LoginRequiredMixin, View):
+    redirect_field_name = 'redirect_to'
+    login_url = '/login/'
+
+    def get(self, request):
+        return render(request, 'build/index.html', {})
+
+
+@require_POST
+def login_view(request):
+    data = json.loads(request.body)
+    username = data.get('username')
+    password = data.get('password')
+
+    if username is None or password is None:
+        return JsonResponse({'detail': 'Please provide username and password.'}, status=status.HTTP_400_BAD_REQUEST)
+
+    user = authenticate(username=username, password=password)
+
+    if user is None:
+        return JsonResponse({'detail': 'Invalid credentials.'}, status=status.HTTP_400_BAD_REQUEST)
+
+    login(request, user)
+    return JsonResponse({'detail': 'Successfully logged in.'})
+
+
+def logout_view(request):
+    if not request.user.is_authenticated:
+        return JsonResponse({'detail': 'You\'re not logged in.'}, status=status.HTTP_400_BAD_REQUEST)
+
+    logout(request)
+    return JsonResponse({'detail': 'Successfully logged out.'})
+
+
+@ensure_csrf_cookie
+def session_view(request):
+    if not request.user.is_authenticated:
+        return JsonResponse({'isAuthenticated': False})
+
+    return JsonResponse({'isAuthenticated': True})
